Move types from JaiaContext to types/context-types
