+++ src/web/containers/BotDetails/BotDetails.tsx
@@ -3,12 +3,9 @@ import React, { useContext, useEffect } from "react";
 // Jaia Imports
 import HealthRow from "../../components/HealthRow/HealthRow";
 import DeleteMissionButton from "../../components/DeleteMissionButton/DeleteMissionButton";
-import {
-    JaiaContext,
-    JaiaContextType,
-    JaiaDispatchContext,
-    JaiaAction,
-} from "../../context/JaiaContext";
+import { JaiaContext, JaiaDispatchContext } from "../../context/JaiaContext";
+import { JaiaAction } from "../../types/context-types";
+import { JaiaContextType } from "../../types/context-types";
 import { JaiaActions } from "../../context/jaia-actions";
 
 import StopButton from "../../components/StopButton/StopButton";
